clip-generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.12
pydantic[dotenv]==1.10.12
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.0
aiofiles==23.2.0
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
librosa==0.10.1
ffmpeg-python==0.2.0
